requirements.txt
mlflow[genai]
agno
python-dotenv
langgraph
langchain-google-genai
langchain
opentelemetry-exporter-otlp openinference-instrumentation-agno
fastapi
uvicorn
